Feat: config api 경로 추가
